ACE-Step-main/requirements.txt
datasets==3.4.1
diffusers>=0.33.0
gradio
librosa==0.11.0
loguru==0.7.3
matplotlib==3.10.1
numpy
pypinyin==0.53.0
pytorch_lightning==2.5.1
soundfile==0.13.1
torch
torchaudio
torchvision
tqdm
transformers==4.50.0
py3langid==0.3.0
hangul-romanize==0.1.0
num2words==0.5.14
spacy==3.8.4
accelerate==1.6.0
cutlet
fugashi[unidic-lite]
click
peft
tensorboard
tensorboardX

gptqmodel==2.2.0
h5py==3.13.0
hdf5plugin==5.1.0
huggingface-hub==0.32.3
natsort==8.4.0
optimum==1.25.3
prodigyopt==1.1.2
qwen-omni-utils==0.0.4
requests==2.32.3
safetensors==0.5.3
wandb==0.20.1
google-generativeai
